core: move to using browserify to compile jsx/require statements
Works nicely but requires a compile step and more installation up
front.  A learning experience if not a good thing.  Should allow
for easier inclusion of other client-side code in the future without
a CDN.  Possible to generate client app and just include a single
generated JS file into other projects if desired.